requirements.txt
PyYAML>=6.0
python-docx>=1.1.0
openpyxl>=3.1.2
python-pptx>=0.6.23
pypdf>=4.2.0
tqdm>=4.66.0

# Windows shortcut (.lnk) resolution
pywin32>=306; platform_system=="Windows"
